JavaDoc comments added.

+++ src/Handlers/GUI_Handler.java
@@ -10,7 +10,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import GUI.GUI_Elements;
-import GUI.SProjectInformation;
 import GUI.SoftvoperMain;
 import LOCSystem.Developers;
 import LOCSystem.LOC;
@@ -21,6 +20,10 @@ import SaveSystem.Save;
 import SaveSystem.SaveSystem;
 import com.fasterxml.jackson.core.JsonProcessingException;
 
+/**
+ * Handles the various GUI actions and transitions between different screens and operations in the game.
+ * It processes button clicks related to saving, loading, creating games, and navigating between different panels.
+ */
 public class GUI_Handler implements ActionListener
 {
     private Save currentSave;
@@ -30,6 +33,13 @@ public class GUI_Handler implements ActionListener
     {
         this.currentSave = save;
     }
+
+    /**
+     * This method handles the actions triggered by button clicks in the GUI.
+     * It processes the action commands for switching panels, loading and saving games, and navigating menus.
+     *
+     * @param event the action event triggered by the button click.
+     */
     public void actionPerformed(ActionEvent event) {
         String action = event.getActionCommand();
         switch (action)
@@ -58,10 +68,12 @@ public class GUI_Handler implements ActionListener
                 GUI_Elements.InitializeInputFrame();
                 break;
             case "CreateGame_Old":
+                // Load an existing game
                 GUI_Elements.window.getContentPane().removeAll();
                 String saveJson = SaveSystem.GetSave(currentSave._id);
                 String[] contents = SaveSystem.ParseJsonStringOneSave(saveJson);
 
+                // Use regex to extract developer, supporter, and project data from the save file
                 Pattern pattern = Pattern.compile("\\{[^}]*\\}");
                 Matcher matcherD = pattern.matcher(contents[4]);
                 Matcher matcherS = pattern.matcher(contents[5]);
@@ -71,6 +83,7 @@ public class GUI_Handler implements ActionListener
                 ArrayList<String> supporter_strings = new ArrayList<String>();
                 ArrayList<String> project_strings = new ArrayList<String>();
 
+                // Collect developer, supporter, and project data
                 while (matcherD.find())
                 {
                     developer_strings.add(matcherD.group());
@@ -84,17 +97,21 @@ public class GUI_Handler implements ActionListener
                     project_strings.add(matcherP.group());
                 }
 
+                // Create developers, supporters, and projects from the saved data
                 ArrayList<Developers> developers = LOC.CreateSavedDevelopers(developer_strings);
                 ArrayList<Supporter> supporter = LOC.CreateSavedSupporters(supporter_strings);
                 ArrayList<SProject> project = LOC.CreateSavedProjects(project_strings);
+
+                // Load the save data into the game state
                 LOC.loc_cnt = Float.parseFloat(contents[2]);
-                SCoin.SCoin_count = Integer.parseInt(contents[3]);
-                SaveSystem.instant_save = new Save(contents[0], contents[1], Float.parseFloat(contents[2]), Integer.parseInt(contents[3]), developers, supporter, project);
-                SoftvoperMain.CreateGameMenu();
-                SCoin.DevelopSavedApp(SaveSystem.instant_save.project);
+                SCoin.SCoin_count = Float.parseFloat(contents[3]);
+                SaveSystem.instant_save = new Save(contents[0], contents[1], Float.parseFloat(contents[2]), Float.parseFloat(contents[3]), developers, supporter, project);
+
+                SoftvoperMain.CreateGameMenu();  // Create the game menu
+                SCoin.DevelopSavedApp(SaveSystem.instant_save.project);  // Start developing saved app projects
                 GUI_Elements.window.revalidate();
                 GUI_Elements.window.repaint();
-                LOC.UpdateLOC();
+                LOC.UpdateLOC();  // Update the LOC (lines of code) tracker
                 break;
             case "Save":
                 try
